backend/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
flake8==3.8.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
